Miscellaneous changes + Icon + Bug fixes
Location automatically detected when you make a new post -> erstwhile button removed
Sign out, sign in and username changing feature completed.
App icon added
Automatic re-login on logout bug squashed
The location on a post is now clickable and takes user to
Google Maps (to the app, or if not available, then to the
website).